requirements.txt.txt
streamlit
pandas
plotly-express
openpyxl